+++ CHANGELOG.md
@@ -7,6 +7,7 @@ yyyy/mm/dd Version x.y.z
 - Renamed `crystalfontz atx` arg `--power-pulse-length-seconds` to `--power-pulse-length`
 - Improved CLI help text
 - Add `--for SECONDS` option to `crystalfontz listen` and `crystalfontz effects` that closes the commands after a certain amount of time
+- Renamed `client` contextmanager to `connection`
 
 2025/01/06 Version 2.0.0
 ------------------------